Allow signer to compile to WASM while also allowing EIP712 primatives to be used in WASM
